+++ packages/terra-progress-bar/src/terra-dev-site/doc/example/ProgressBarColor.jsx
@@ -5,13 +5,13 @@ import './colors.css';
 
 const ProgressBarColor = () => (
   <div>
-    <ProgressBar valueText="5%" value={5} max={10} className="orange-bar" />
+    <ProgressBar valueText="5%" value={5} max={10} colorClass="orange-bar" />
     <br />
     <br />
-    <ProgressBar valueText="7.5%" value={7.5} max={10} className="green-bar" />
+    <ProgressBar valueText="7.5%" value={7.5} max={10} colorClass="green-bar" />
     <br />
     <br />
-    <ProgressBar valueText="10%" value={10} max={10} className="red-bar" />
+    <ProgressBar valueText="10%" value={10} max={10} colorClass="red-bar" />
   </div>
 );