+++ packages/terra-progress-bar/src/terra-dev-site/doc/example/ProgressBarSize.jsx
@@ -5,19 +5,19 @@ import './colors.css';
 
 const ProgressBarSize = () => (
   <div>
-    <ProgressBar heightSize="tiny" valueText="15%" value={15} className="green-bar" />
+    <ProgressBar heightSize="tiny" valueText="15%" value={15} />
     <br />
     <br />
-    <ProgressBar heightSize="small" valueText="30%" value={30} className="green-bar" />
+    <ProgressBar heightSize="small" valueText="30%" value={30} />
     <br />
     <br />
-    <ProgressBar heightSize="medium" valueText="45%" value={45} className="green-bar" />
+    <ProgressBar heightSize="medium" valueText="45%" value={45} />
     <br />
     <br />
-    <ProgressBar heightSize="large" valueText="60%" value={60} className="green-bar" />
+    <ProgressBar heightSize="large" valueText="60%" value={60} />
     <br />
     <br />
-    <ProgressBar heightSize="huge" valueText="75%" value={75} className="green-bar" />
+    <ProgressBar heightSize="huge" valueText="75%" value={75} />
   </div>
 );