+++ packages/terra-progress-bar/src/terra-dev-site/doc/example/ProgressBarTwoColors.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
+import ProgressBar from 'terra-progress-bar/lib/ProgressBar';
+import './colors.css';
+
+const ProgressBarGradient = () => (
+  <div>
+    <ProgressBar valueText="5%" value={5} max={10} colorClass="green-bar-light-grey-background" />
+    <br />
+    <br />
+    <ProgressBar valueText="3%" value={3} max={10} colorClass="red-bar-black-background" />
+    <br />
+    <br />
+    <ProgressBar valueText="8%" value={8} max={10} colorClass="orange-bar-dark-blue-background" />
+  </div>
+);
+
+export default ProgressBarGradient;