+++ packages/terra-progress-bar/src/terra-dev-site/test/progress-bar/ProgressBarColor.test.jsx
@@ -4,13 +4,13 @@ import '../../doc/example/colors.css';
 
 const ProgressBarColor = () => (
   <div>
-    <ProgressBar id="progressbarWithNamedColor" valueText="5%" value={5} max={10} className="orange-bar" />
+    <ProgressBar id="progressbarWithNamedColor" valueText="5%" value={5} max={10} colorClass="orange-bar" />
     <br />
     <br />
-    <ProgressBar id="progressbarWithHexColor" valueText="7.5%" value={7.5} max={10} className="green-bar" />
+    <ProgressBar id="progressbarWithHexColor" valueText="7.5%" value={7.5} max={10} colorClass="green-bar" />
     <br />
     <br />
-    <ProgressBar id="progressbarWithRGBColor" valueText="10%" value={10} max={10} className="red-bar" />
+    <ProgressBar id="progressbarWithRGBColor" valueText="10%" value={10} max={10} colorClass="red-bar" />
   </div>
 );