+++ packages/terra-progress-bar/src/terra-dev-site/test/progress-bar/ProgressBarDefault.test.jsx
@@ -3,7 +3,7 @@ import ProgressBar from '../../../ProgressBar';
 import '../../doc/example/colors.css';
 
 const ProgressBarDefault = () => (
-  <ProgressBar id="progressbar" valueText="15%" value={15} className="yellow-bar" />
+  <ProgressBar id="progressbar" valueText="15%" value={15} />
 );
 
 export default ProgressBarDefault;