+++ packages/terra-progress-bar/tests/jest/ProgressBar.test.jsx
@@ -17,6 +17,11 @@ it('should render a ProgressBar component with large heightSize and 60% fill', (
   expect(wrapper).toMatchSnapshot();
 });
 
+it('should render a ProgressBar component with default heightSize and 50% fill and colorClass prop', () => {
+  const wrapper = shallow(<ProgressBar value={60} valueText="60%" max={120} colorClass="yellow-bar" />);
+  expect(wrapper).toMatchSnapshot();
+});
+
 it('should render a ProgressBar component with default heightSize 50% fill and custom props and style', () => {
   const wrapper = shallow(<ProgressBar value={6} valueText="6%" max={12} dir="rtl" />);
   expect(wrapper).toMatchSnapshot();
@@ -26,7 +31,7 @@ it('should render a ProgressBar component with default heightSize 50% fill and c
 describe('Default progress bar', () => {
   it('should have classes progressbar, small and zero fill value', () => {
     const wrapper = shallow(<ProgressBar value={0} valueText="0%" />);
-    expect(wrapper.prop('className')).toContain('progress-bar small');
+    expect(wrapper.prop('className')).toContain('progress-bar small default-color');
     expect(wrapper.prop('value')).toEqual(0);
   });
 });
@@ -46,6 +51,15 @@ describe('Progress bar with default heightSize prop', () => {
     const wrapper = shallow(<ProgressBar value={10} valueText="10%" max={10} />);
     expect(wrapper.prop('value')).toEqual(100);
     expect(wrapper.prop('max')).toEqual(100);
+  });
+});
+
+describe('Progress bar with colorClass prop', () => {
+  it('value=3, max=10; should have class yellow-bar, small and 30% fill value', () => {
+    const wrapper = shallow(<ProgressBar title="ProgressBarTest" value={3} valueText="30%" max={10} colorClass="yellow-bar" />);
+    expect(wrapper.prop('value')).toEqual(30);
+    expect(wrapper.prop('max')).toEqual(100);
+    expect(wrapper.prop('className')).toContain('yellow-bar');
   });
 });